Refactored VeggieNote and AirSlicer scripts
-Moved functions for spawning, deleting, hit, and misses to AirSlicer
-VeggieNote script is now responsible for just Dotween path
