+++ Assets/Scenes/Minigames/Air Slicer.unity
@@ -431,6 +431,50 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d0d5426b18d640c4db8c032c2391b126, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!1 &1026153491
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1026153493}
+  - component: {fileID: 1026153492}
+  m_Layer: 0
+  m_Name: Collision
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!65 &1026153492
+BoxCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1026153491}
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_Enabled: 1
+  serializedVersion: 2
+  m_Size: {x: 1, y: 1, z: 1}
+  m_Center: {x: 0, y: 0, z: 0}
+--- !u!4 &1026153493
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1026153491}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0.969, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 8
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1138025222
 GameObject:
   m_ObjectHideFlags: 0
@@ -888,6 +932,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 9e064eaea7b006d4e9852fdb054f433b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  cubeTransform: {fileID: 1139206299447070057, guid: b04b7d2d14b91144286fcd9184117061,
+    type: 3}
+  cubeMaterial: {fileID: 10303, guid: 0000000000000000f000000000000000, type: 0}
   prefab:
   - {fileID: 1139206299447070060, guid: b04b7d2d14b91144286fcd9184117061, type: 3}
   - {fileID: 1139206299447070060, guid: b04b7d2d14b91144286fcd9184117061, type: 3}
