+++ Assets/Scripts/Air Slicer/AirSlicer.cs
@@ -1,27 +1,38 @@
 ﻿using UnityEngine;
 using System.Collections;
 using System.Collections.Generic;
+using DG.Tweening;
 
 public class AirSlicer : MonoBehaviour
 {
+    public Transform cubeTransform;
+    public Material cubeMaterial;
     public GameObject[] prefab;
+    //public GameObject currentVeggieNote;
     public static float animationDuration;
-    private GameObject currentVeggieNote;
-
-    //private List<GameObject> activeVeggieNotes;
-    private Queue<GameObject> activeVeggieNotes = new Queue<GameObject>();
 
     [FMODUnity.EventRef]
     public string[] sfx;
 
     void Start()
     {
-        //activeVeggieNotes = new List<GameObject>();
+        GameManager.ValidHit += SuccessfulHit;
+        GameManager.MissedHit += UnSuccessfulHit;
+   
         BeatSystem.OnMarker += SpawnNote;
     }
 
+    void Update()
+    {
+        if (cubeTransform.position.y > 0.95)
+        {
+          StartCoroutine("DeleteNote");
+        }
+    }
     private void OnDisable()
     {
+        GameManager.ValidHit -= SuccessfulHit;
+        GameManager.MissedHit -= UnSuccessfulHit;
         BeatSystem.OnMarker -= SpawnNote;
 
     }
@@ -39,19 +50,7 @@ public class AirSlicer : MonoBehaviour
 
     private void SpawnOnion()
     {
-        // --------------------- Queue Method ----------------------------------//
-        currentVeggieNote = Instantiate(prefab[0], new Vector3(-4.9f, -5.03f, 0f), Quaternion.identity);
-        activeVeggieNotes.Enqueue(currentVeggieNote);
-        Debug.Log("Queue Size:" + activeVeggieNotes.Count);
-        
-        // ---------------------- List Method -----------------------------------//
-        //GameObject newVeggieNote = Instantiate(prefab[0], new Vector3(-5.32f, -6.03f, 0f), Quaternion.identity);
-        //activeVeggieNotes.Add(newVeggieNote);
-
-
-        // ---------------------- Current Method (deletes all clones) -----------//
-        //Instantiate(prefab[0], new Vector3(-5.32f, -6.03f, 0f), Quaternion.identity);
-        //Destroy(Instantiate(prefab[0], new Vector3(-5.32f, -6.03f, 0f), Quaternion.identity), 1.5f);
+        Instantiate(prefab[0], new Vector3(-4.9f, -5.03f, 0f), Quaternion.identity);
     }
     private void SpawnCarrot()
     {
@@ -63,17 +62,48 @@ public class AirSlicer : MonoBehaviour
         Instantiate(prefab[2], new Vector3(-5.32f, -6.03f, 0f), Quaternion.identity);
     }
 
-    public void DestroyVeggieNote()
+    private void SuccessfulHit()
     {
-        if (activeVeggieNotes.Count > 0)
-        {
-            Debug.Log("Total Notes: " + activeVeggieNotes.Count);
-            // VeggieNote is popped from Queue and then Destroyed
-            GameObject poppedVeggieNote = activeVeggieNotes.Dequeue();
-            //poppedVeggieNote.SetActive(false);
-            Destroy(poppedVeggieNote, 1.25f);
+        //if (note != null)
+        //{
+        cubeMaterial.DOColor(Color.green, 1);
+        GameObject currentVeggieNote = GameObject.Find("VeggieNote(Clone)");
+        Destroy(Instantiate(prefab[3], new Vector3(currentVeggieNote.transform.position.x, currentVeggieNote.transform.position.y, 0f), Quaternion.identity), 1f);
+        DOTween.Kill(currentVeggieNote);
+        Destroy(currentVeggieNote);
 
-            //Destroy(activeVeggieNotes[0]);
-        }
+            //Destroy(currentVeggieNote);
+        Debug.Log("This should have removed the oldest clone");
+
+            //Debug.Log("Making a function call to DestroyVeggieNote()");
+            //AirSlicer.DestroyVeggieNote();
+            //Destroy(Instantiate(this.cutVeggie, new Vector3(transform.position.x, transform.position.y, 0f), Quaternion.identity), 2f);
+            //Instantiate(cutVeggie, new Vector3(transform.position.x, transform.position.y, 0f), Quaternion.identity);     
+       // }
     }
+
+    private void UnSuccessfulHit()
+    {
+        //if (note != null)
+        //{
+            cubeMaterial.DOColor(Color.red, 1);
+            GameObject currentVeggieNote = GameObject.Find("VeggieNote(Clone)");
+            //Instantiate(cutVeggie, new Vector3(currentVeggieNote.transform.position.x, currentVeggieNote.transform.position.y, 0f), Quaternion.identity);
+            Destroy(currentVeggieNote);
+       // }
+    }
+
+    void OnTriggerEnter(Collider other)
+    {
+        Debug.Log("Collision!!!!");
+        DOTween.Kill(other.gameObject);
+        Destroy(other.gameObject);
+    }
+
+    // Waits a bit to ensure cube has reached it's final destination before being deleted, otherwise generates DOTween warnings. 
+    IEnumerator DeleteNote()
+     {
+       yield return new WaitForSeconds(0.1f);
+       //Destroy(note);
+     }
 }
